Initial commit. Where I'm at today.
To anyone who may read this, that isn't me: A lot of this was written more or less stream-of-conciousness style as I was doing some related reading. At this moment, I'm not sure what will or will not make the final draft. I also have a tendency to leave comments for myself in the middle of paragraphs. Enjoy.